refactor(): move classes into a common package
