+++ src/main/java/com/axway/gw/es/yaml/model/entity/EntityDTO.java
@@ -1,6 +1,6 @@
-package com.axway.gw.es.model.entity;
+package com.axway.gw.es.yaml.model.entity;
 
-import com.axway.gw.es.model.type.FieldDTO;
+import com.axway.gw.es.yaml.model.type.FieldDTO;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import java.util.LinkedHashMap;
