+++ src/main/java/com/axway/gw/es/yaml/tools/CloneES.java
@@ -1,4 +1,4 @@
-package com.axway.gw.es.tools;
+package com.axway.gw.es.yaml.tools;
 
 import java.util.ArrayList;
 import java.util.Collection;
