+++ src/main/java/com/axway/gw/es/yaml/tools/ConvertToYamlStore.java
@@ -1,4 +1,4 @@
-package com.axway.gw.es.tools;
+package com.axway.gw.es.yaml.tools;
 
 import java.io.File;
 import java.io.IOException;
