+++ src/main/java/com/axway/gw/es/yaml/tools/CreateYamlEntityStore.java
@@ -1,4 +1,4 @@
-package com.axway.gw.es.tools;
+package com.axway.gw.es.yaml.tools;
 
 import com.axway.gw.es.yaml.YamlEntityStore;
